feat: update ssr in home
